+++ gspread/models.py
@@ -1,85 +1,78 @@
 # -*- coding: utf-8 -*-
 
 """
-gspread.models
-~~~~~~~~~~~~~~
+gspread.v4.models
+~~~~~~~~~~~~~~~~~
 
-This module contains common spreadsheets' models
+This module contains common spreadsheets' models.
 
 """
 
-from collections import defaultdict
-from itertools import chain
+try:
+    from urllib.parse import quote
+except:
+    from urllib import quote
 
-from xml.etree import ElementTree
-from xml.etree.ElementTree import Element, SubElement
+from ..base import BaseCell, BaseSpreadsheet
 
-from . import urlencode
-from .ns import _ns, _ns1, ATOM_NS, BATCH_NS, SPREADSHEET_NS
-from .urls import construct_url
+from ..exceptions import WorksheetNotFound, CellNotFound
 
-from .utils import (
-    finditem,
-    numericise_all,
-    rowcol_to_a1,
+from ..utils import (
     a1_to_rowcol,
-    wid_to_gid,
-    cast_to_a1_notation
+    rowcol_to_a1,
+    cast_to_a1_notation,
+    numericise_all,
+    finditem
 )
 
-from .exceptions import WorksheetNotFound, CellNotFound
+from .utils import fill_gaps, cell_list_to_rect
 
-from .base import BaseSpreadsheet, BaseCell
+from .urls import (
+    SPREADSHEET_URL,
+    SPREADSHEET_VALUES_URL,
+    SPREADSHEET_BATCH_UPDATE_URL,
+    SPREADSHEET_VALUES_APPEND_URL,
+    SPREADSHEET_VALUES_CLEAR_URL
+)
 
 try:
     unicode
 except NameError:
     basestring = unicode = str
 
-# Patch ElementTree._escape_attrib
-_elementtree_escape_attrib = ElementTree._escape_attrib
-
-
-def _escape_attrib(text, encoding=None, replace=None):
-    try:
-        text = _elementtree_escape_attrib(text)
-    except TypeError as e:
-        if str(e) == '_escape_attrib() takes exactly 2 arguments (1 given)':
-            text = _elementtree_escape_attrib(text, encoding)
-    entities = {'\n': '&#10;', '\r': '&#13;', '\t': '&#9;'}
-    for key, value in entities.items():
-        text = text.replace(key, value)
-    return text
-
-
-ElementTree._escape_attrib = _escape_attrib
-
 
 class Spreadsheet(BaseSpreadsheet):
-    """ A class for a spreadsheet object."""
-
-    def __init__(self, client, feed_entry):
+    """The class that represents a spreadsheet."""
+    def __init__(self, client, properties):
         self.client = client
-        self._sheet_list = []
-        self._feed_entry = feed_entry
-        self._id = feed_entry.find(_ns('id')).text.split('/')[-1]
-        self._title = feed_entry.find(_ns('title')).text
-        self._updated = feed_entry.find(_ns('updated')).text
+        self._properties = properties
 
     @property
     def id(self):
         """Spreadsheet ID."""
-        return self._id
+        return self._properties['id']
 
     @property
     def title(self):
         """Spreadsheet title."""
-        return self._title
+        try:
+            return self._properties['title']
+        except KeyError:
+            metadata = self.fetch_sheet_metadata()
+            self._properties.update(metadata['properties'])
+            return self._properties['title']
 
     @property
     def updated(self):
-        """Updated time in RFC 3339 format"""
-        return self._updated
+        """.. deprecated:: 2.0
+        This feature is not supported in Sheets API v4.
+        """
+        import warnings
+        warnings.warn(
+            "Spreadsheet.updated() is deprecated, "
+            "this feature is not supported in Sheets API v4",
+            DeprecationWarning
+        )
 
     @property
     def sheet1(self):
@@ -95,14 +88,98 @@ class Spreadsheet(BaseSpreadsheet):
                                   repr(self.title),
                                   self.id)
 
-    def get_id_fields(self):
-        return {'spreadsheet_id': self.id}
+    def batch_update(self, body):
+        r = self.client.request(
+            'post',
+            SPREADSHEET_BATCH_UPDATE_URL % self.id,
+            json=body
+        )
 
-    def _fetch_sheets(self):
-        self._sheet_list = []
-        feed = self.client.get_worksheets_feed(self)
-        for elem in feed.findall(_ns('entry')):
-            self._sheet_list.append(Worksheet(self, elem))
+        return r.json()
+
+    def values_append(self, range, params, body):
+        url = SPREADSHEET_VALUES_APPEND_URL % (self.id, quote(range, safe=''))
+        r = self.client.request('post', url, params=params, json=body)
+        return r.json()
+
+    def values_clear(self, range):
+        url = SPREADSHEET_VALUES_CLEAR_URL % (self.id, quote(range, safe=''))
+        r = self.client.request('post', url)
+        return r.json()
+
+    def values_get(self, range, params=None):
+        url = SPREADSHEET_VALUES_URL % (self.id, quote(range, safe=''))
+        r = self.client.request('get', url, params=params)
+        return r.json()
+
+    def values_update(self, range, params=None, body=None):
+        url = SPREADSHEET_VALUES_URL % (self.id, quote(range, safe=''))
+        r = self.client.request('put', url, params=params, json=body)
+        return r.json()
+
+    def fetch_sheet_metadata(self):
+        params = {'includeGridData': 'false'}
+
+        url = SPREADSHEET_URL % self.id
+
+        r = self.client.request('get', url, params=params)
+
+        return r.json()
+
+    def get_worksheet(self, index):
+        """Returns a worksheet with specified `index`.
+
+        :param index: An index of a worksheet. Indexes start from zero.
+
+        :returns: an instance of :class:`gsperad.v4.models.Worksheet`
+                  or `None` if the worksheet is not found.
+
+        Example. To get first worksheet of a spreadsheet:
+
+        >>> sht = client.open('My fancy spreadsheet')
+        >>> worksheet = sht.get_worksheet(0)
+
+        """
+        sheet_data = self.fetch_sheet_metadata()
+
+        try:
+            properties = sheet_data['sheets'][index]['properties']
+            return Worksheet(self, properties)
+        except (KeyError, IndexError):
+            return None
+
+    def worksheets(self):
+        """Returns a list of all :class:`worksheets <gsperad.v4.models.Worksheet>`
+        in a spreadsheet.
+
+        """
+        sheet_data = self.fetch_sheet_metadata()
+        return [Worksheet(self, x['properties']) for x in sheet_data['sheets']]
+
+    def worksheet(self, title):
+        """Returns a worksheet with specified `title`.
+
+        :param title: A title of a worksheet. If there're multiple
+                      worksheets with the same title, first one will
+                      be returned.
+
+        :returns: an instance of :class:`gsperad.v4.models.Worksheet`.
+
+        Example. Getting worksheet named 'Annual bonuses'
+
+        >>> sht = client.open('Sample one')
+        >>> worksheet = sht.worksheet('Annual bonuses')
+
+        """
+        sheet_data = self.fetch_sheet_metadata()
+        try:
+            item = finditem(
+                lambda x: x['properties']['title'] == title,
+                sheet_data['sheets']
+            )
+            return Worksheet(self, item['properties'])
+        except (StopIteration, KeyError):
+            raise WorksheetNotFound(title)
 
     def add_worksheet(self, title, rows, cols):
         """Adds a new worksheet to a spreadsheet.
@@ -111,20 +188,28 @@ class Spreadsheet(BaseSpreadsheet):
         :param rows: Number of rows.
         :param cols: Number of columns.
 
-        :returns: a newly created :class:`worksheets <Worksheet>`.
+        :returns: a newly created :class:`worksheets <gsperad.v4.models.Worksheet>`.
         """
-        feed = Element('entry', {'xmlns': ATOM_NS,
-                                 'xmlns:gs': SPREADSHEET_NS})
+        body = {
+            'requests': [{
+                'addSheet': {
+                    'properties': {
+                        'title': title,
+                        'sheetType': 'GRID',
+                        'gridProperties': {
+                            'rowCount': rows,
+                            'columnCount': cols
+                        }
+                    }
+                }
+            }]
+        }
 
-        SubElement(feed, 'title').text = title
-        SubElement(feed, 'gs:rowCount').text = str(rows)
-        SubElement(feed, 'gs:colCount').text = str(cols)
+        data = self.batch_update(body)
 
-        url = construct_url('worksheets', self)
-        elem = self.client.post_feed(url, ElementTree.tostring(feed))
+        properties = data['replies'][0]['addSheet']['properties']
 
-        worksheet = Worksheet(self, elem)
-        self._sheet_list.append(worksheet)
+        worksheet = Worksheet(self, properties)
 
         return worksheet
 
@@ -134,80 +219,25 @@ class Spreadsheet(BaseSpreadsheet):
         :param worksheet: The worksheet to be deleted.
 
         """
-        self.client.del_worksheet(worksheet)
-        self._sheet_list.remove(worksheet)
+        body = {
+            'requests': [{
+                'deleteSheet': {'sheetId': worksheet._properties['sheetId']}
+            }]
+        }
 
-    def worksheets(self):
-        """Returns a list of all :class:`worksheets <Worksheet>`
-        in a spreadsheet.
-
-        """
-        if not self._sheet_list:
-            self._fetch_sheets()
-        return self._sheet_list[:]
-
-    def worksheet(self, title):
-        """Returns a worksheet with specified `title`.
-
-        :param title: A title of a worksheet. If there're multiple
-                      worksheets with the same title, first one will
-                      be returned.
-
-        :returns: an instance of :class:`Worksheet`.
-
-        Example. Getting worksheet named 'Annual bonuses'
-
-        >>> sht = client.open('Sample one')
-        >>> worksheet = sht.worksheet('Annual bonuses')
-
-        """
-        if not self._sheet_list:
-            self._fetch_sheets()
-
-        try:
-            return finditem(lambda x: x.title == title, self._sheet_list)
-        except StopIteration:
-            raise WorksheetNotFound(title)
-
-    def get_worksheet(self, index):
-        """Returns a worksheet with specified `index`.
-
-        :param index: An index of a worksheet. Indexes start from zero.
-
-        :returns: an instance of :class:`Worksheet`
-                  or `None` if the worksheet is not found.
-
-        Example. To get first worksheet of a spreadsheet:
-
-        >>> sht = client.open('My fancy spreadsheet')
-        >>> worksheet = sht.get_worksheet(0)
-
-        """
-        if not self._sheet_list:
-            self._fetch_sheets()
-        try:
-            return self._sheet_list[index]
-        except IndexError:
-            return None
+        return self.batch_update(body)
 
 
 class Worksheet(object):
+    """The class that represents a single sheet in a spreadsheet
+    (aka "worksheet").
 
-    """A class for worksheet object."""
+    """
 
-    def __init__(self, spreadsheet, element):
+    def __init__(self, spreadsheet, properties):
         self.spreadsheet = spreadsheet
         self.client = spreadsheet.client
-        self._id = element.find(_ns('id')).text.split('/')[-1]
-        self._title = element.find(_ns('title')).text
-        self._element = element
-
-        try:
-            self.version = self._get_link(
-                'edit', element).get('href').split('/')[-1]
-        except:
-            # not relevant for read-only spreadsheets
-            self.version = None
+        self._properties = properties
 
     def __repr__(self):
         return '<%s %s id:%s>' % (self.__class__.__name__,
@@ -217,83 +247,45 @@ class Worksheet(object):
     @property
     def id(self):
         """Id of a worksheet."""
-        return self._id
+        return self._properties['sheetId']
 
     @property
     def title(self):
         """Title of a worksheet."""
-        return self._title
-
-    @property
-    def gid(self):
-        """Gid of a worksheet."""
-        return wid_to_gid(self._id)
-
-    @property
-    def row_count(self):
-        """Number of rows"""
-        return int(self._element.find(_ns1('rowCount')).text)
-
-    @property
-    def col_count(self):
-        """Number of columns"""
-        return int(self._element.find(_ns1('colCount')).text)
+        return self._properties['title']
 
     @property
     def updated(self):
-        """Updated time in RFC 3339 format"""
-        return self._element.find(_ns('updated')).text
-
-    def get_id_fields(self):
-        return {'spreadsheet_id': self.spreadsheet.id,
-                'worksheet_id': self.id}
-
-    def _cell_addr(self, row, col):
-        return 'R%sC%s' % (row, col)
-
-    def _get_link(self, link_type, feed):
-        return finditem(lambda x: x.get('rel') == link_type,
-                        feed.findall(_ns('link')))
-
-    def _fetch_cells(self):
-        feed = self.client.get_cells_feed(self)
-        return [Cell(self, elem) for elem in feed.findall(_ns('entry'))]
-
-    def get_int_addr(self, label):
-        """Translates cell's label address to a tuple of integers.
-
-        .. deprecated:: 0.5
-           Use :func:`utils.a1_to_rowcol` instead.
-
+        """.. deprecated:: 2.0
+        This feature is not supported in Sheets API v4.
         """
         import warnings
         warnings.warn(
-            "Worksheet.get_int_addr() is deprecated, "
-            "use utils.a1_to_rowcol() instead",
+            "Worksheet.updated() is deprecated, "
+            "this feature is not supported in Sheets API v4",
             DeprecationWarning
         )
-        return a1_to_rowcol(label)
 
-    def get_addr_int(self, row, col):
-        """Translates cell's tuple of integers to a cell label.
+    @property
+    def row_count(self):
+        """Number of rows."""
+        return self._properties['gridProperties']['rowCount']
 
-        .. deprecated:: 0.5
-           Use :func:`utils.rowcol_to_a1` instead.
+    @property
+    def col_count(self):
+        """Number of columns."""
+        return self._properties['gridProperties']['columnCount']
 
-        """
-        import warnings
-        warnings.warn(
-            "Worksheet.get_addr_int() is deprecated, "
-            "use utils.rowcol_to_a1() instead",
-            DeprecationWarning
-        )
-        return rowcol_to_a1(row, col)
-
-    def acell(self, label):
-        """Returns an instance of a :class:`Cell`.
+    def acell(self, label, value_render_option='FORMATTED_VALUE'):
+        """Returns an instance of a :class:`gspread.v4.models.Cell`.
 
         :param label: String with cell label in common format, e.g. 'B1'.
                       Letter case is ignored.
+        :param value_render_option: Determines how values should be rendered
+                                    in the the output. See `ValueRenderOption`_
+                                    in the Sheets API.
+
+        .. _ValueRenderOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueRenderOption
 
         Example:
 
@@ -301,14 +293,23 @@ class Worksheet(object):
         <Cell R1C1 "I'm cell A1">
 
         """
-        return self.cell(*(a1_to_rowcol(label)))
 
-    def cell(self, row, col):
-        """Returns an instance of a :class:`Cell` positioned in `row`
-           and `col` column.
+        return self.cell(
+            *(a1_to_rowcol(label)),
+            value_render_option=value_render_option
+        )
+
+    def cell(self, row, col, value_render_option='FORMATTED_VALUE'):
+        """Returns an instance of a :class:`gspread.v4.models.Cell` positioned
+        in `row` and `col` column.
 
         :param row: Integer row number.
         :param col: Integer column number.
+        :param value_render_option: Determines how values should be rendered
+                                    in the the output. See `ValueRenderOption`_
+                                    in the Sheets API.
+
+        .. _ValueRenderOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueRenderOption
 
         Example:
 
@@ -316,9 +317,19 @@ class Worksheet(object):
         <Cell R1C1 "I'm cell A1">
 
         """
-        feed = self.client.get_cells_cell_id_feed(self,
-                                                  self._cell_addr(row, col))
-        return Cell(self, feed)
+
+        range_label = '%s!%s' % (self.title, rowcol_to_a1(row, col))
+        data = self.spreadsheet.values_get(
+            range_label,
+            params={'valueRenderOption': value_render_option}
+        )
+
+        try:
+            value = data['values'][0][0]
+        except KeyError:
+            value = ''
+
+        return Cell(row, col, value)
 
     @cast_to_a1_notation
     def range(self, name):
@@ -345,51 +356,58 @@ class Worksheet(object):
             [<Cell R1C1 "42">, ...]
 
         """
-        feed = self.client.get_cells_feed(
-            self,
-            params={'range': name, 'return-empty': 'true'}
+
+        range_label = '%s!%s' % (self.title, name)
+
+        data = self.spreadsheet.values_get(range_label)
+
+        start, end = name.split(':')
+        (row_offset, column_offset) = a1_to_rowcol(start)
+        (last_row, last_column) = a1_to_rowcol(end)
+
+        values = data.get('values', [])
+
+        rect_values = fill_gaps(
+            values,
+            rows=last_row - row_offset + 1,
+            cols=last_column - column_offset + 1
         )
-        return [Cell(self, elem) for elem in feed.findall(_ns('entry'))]
+
+        return [
+            Cell(row=i + row_offset, col=j + column_offset, value=value)
+            for i, row in enumerate(rect_values)
+            for j, value in enumerate(row)
+        ]
 
     def get_all_values(self):
         """Returns a list of lists containing all cells' values as strings.
+
         """
-        cells = self._fetch_cells()
 
-        # defaultdicts fill in gaps for empty rows/cells not returned by gdocs
-        rows = defaultdict(lambda: defaultdict(str))
-        for cell in cells:
-            row = rows.setdefault(int(cell.row), defaultdict(str))
-            row[cell.col] = cell.value
+        data = self.spreadsheet.values_get(self.title)
 
-        # we return a whole rectangular region worth of cells, including
-        # empties
-        if not rows:
+        try:
+            return fill_gaps(data['values'])
+        except KeyError:
             return []
 
-        all_row_keys = chain.from_iterable(row.keys() for row in rows.values())
-        rect_cols = range(1, max(all_row_keys) + 1)
-        rect_rows = range(1, max(rows.keys()) + 1)
-
-        return [[rows[i][j] for j in rect_cols] for i in rect_rows]
-
     def get_all_records(self, empty2zero=False, head=1, default_blank=""):
-        """Returns a list of dictionaries, all of them having:
-            - the contents of the spreadsheet's with the head row as keys,
-            And each of these dictionaries holding
-            - the contents of subsequent rows of cells as values.
-
+        """Returns a list of dictionaries, all of them having the contents
+        of the spreadsheet with the head row as keys and each of these
+        dictionaries holding the contents of subsequent rows of cells
+        as values.
 
         Cell values are numericised (strings that can be read as ints
         or floats are converted).
 
-        :param empty2zero: determines whether empty cells are converted to zeros.
-        :param head: determines wich row to use as keys, starting from 1
-            following the numeration of the spreadsheet.
-        :param default_blank: determines whether empty cells are converted to
-            something else except empty string or zero.
-
+        :param empty2zero: determines whether empty cells are converted
+                           to zeros.
+        :param head: determines wich row to use as keys, starting
+                     from 1 following the numeration of the spreadsheet.
+        :param default_blank: determines whether empty cells are converted
+                              to something else except empty string or zero.
         """
+
         idx = head - 1
 
         data = self.get_all_values()
@@ -399,97 +417,111 @@ class Worksheet(object):
 
         return [dict(zip(keys, row)) for row in values]
 
-    def row_values(self, row):
+    def row_values(self, row, value_render_option='FORMATTED_VALUE'):
         """Returns a list of all values in a `row`.
 
         Empty cells in this list will be rendered as :const:`None`.
 
+        :param row: Integer row number.
+        :param value_render_option: Determines how values should be rendered
+                                    in the the output. See `ValueRenderOption`_
+                                    in the Sheets API.
+
+        .. _ValueRenderOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueRenderOption
+
         """
-        start_cell = rowcol_to_a1(row, 1)
-        end_cell = rowcol_to_a1(row, self.col_count)
 
-        row_cells = self.range('%s:%s' % (start_cell, end_cell))
-        return [cell.value for cell in row_cells]
+        range_label = '%s!A%s:%s' % (self.title, row, row)
 
-    def col_values(self, col):
+        data = self.spreadsheet.values_get(
+            range_label,
+            params={'valueRenderOption': value_render_option}
+        )
+
+        try:
+            return data['values'][0]
+        except KeyError:
+            return []
+
+    def col_values(self, col, value_render_option='FORMATTED_VALUE'):
         """Returns a list of all values in column `col`.
 
         Empty cells in this list will be rendered as :const:`None`.
 
+        :param col: Integer column number.
+        :param value_render_option: Determines how values should be rendered
+                                    in the the output. See `ValueRenderOption`_
+                                    in the Sheets API.
+
+        .. _ValueRenderOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueRenderOption
+
         """
-        start_cell = rowcol_to_a1(1, col)
-        end_cell = rowcol_to_a1(self.row_count, col)
 
-        row_cells = self.range('%s:%s' % (start_cell, end_cell))
-        return [cell.value for cell in row_cells]
+        start_label = rowcol_to_a1(1, col)
+        range_label = '%s!%s:%s' % (self.title, start_label, start_label[:-1])
 
-    def update_acell(self, label, val):
+        data = self.spreadsheet.values_get(
+            range_label,
+            params={
+                'valueRenderOption': value_render_option,
+                'majorDimension': 'COLUMNS'
+            }
+        )
+
+        try:
+            return data['values'][0]
+        except KeyError:
+            return []
+
+    def update_acell(self, label, value):
         """Sets the new value to a cell.
 
         :param label: String with cell label in common format, e.g. 'B1'.
                       Letter case is ignored.
-        :param val: New value.
+        :param value: New value.
 
-        Example:
+        Example::
 
-            worksheet.update_acell('A1', '42') # this could be 'a1' as well
+            worksheet.update_acell('A1', '42')
 
         """
-        return self.update_cell(*(a1_to_rowcol(label)), val=val)
+        return self.update_cell(*(a1_to_rowcol(label)), value=value)
 
-    def update_cell(self, row, col, val):
+    def update_cell(self, row, col, value):
         """Sets the new value to a cell.
 
         :param row: Row number.
         :param col: Column number.
-        :param val: New value.
+        :param value: New value.
 
         Example::
 
             worksheet.update_cell(1, 1, '42')
 
         """
-        feed = self.client.get_cells_cell_id_feed(self,
-                                                  self._cell_addr(row, col))
-        cell_elem = feed.find(_ns1('cell'))
-        cell_elem.set('inputValue', unicode(val))
-        uri = self._get_link('edit', feed).get('href')
+        range_label = '%s!%s' % (self.title, rowcol_to_a1(row, col))
 
-        self.client.put_feed(uri, ElementTree.tostring(feed))
+        data = self.spreadsheet.values_update(
+            range_label,
+            params={
+                'valueInputOption': 'USER_ENTERED'
+            },
+            body={
+                'values': [[value]]
+            }
+        )
 
-    def _create_update_feed(self, cell_list):
-        feed = Element('feed', {'xmlns': ATOM_NS,
-                                'xmlns:batch': BATCH_NS,
-                                'xmlns:gs': SPREADSHEET_NS})
+        return data
 
-        id_elem = SubElement(feed, 'id')
-
-        id_elem.text = construct_url('cells', self)
-
-        for cell in cell_list:
-            entry = SubElement(feed, 'entry')
-
-            SubElement(entry, 'batch:id').text = cell.element.find(
-                _ns('title')).text
-            SubElement(entry, 'batch:operation', {'type': 'update'})
-            SubElement(entry, 'id').text = cell.element.find(_ns('id')).text
-
-            edit_link = finditem(lambda x: x.get('rel') == 'edit',
-                                 cell.element.findall(_ns('link')))
-
-            SubElement(entry, 'link', {'rel': 'edit',
-                                       'type': edit_link.get('type'),
-                                       'href': edit_link.get('href')})
-
-            SubElement(entry, 'gs:cell', {'row': str(cell.row),
-                                          'col': str(cell.col),
-                                          'inputValue': unicode(cell.value)})
-        return feed
-
-    def update_cells(self, cell_list):
+    def update_cells(self, cell_list, value_input_option='RAW'):
         """Updates cells in batch.
 
         :param cell_list: List of a :class:`Cell` objects to update.
+        :param value_input_option: Determines how input data should be
+                                   interpreted. See `ValueInputOption`_
+                                   in the Sheets API.
+
+        .. _ValueInputOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueInputOption
 
         Example::
 
@@ -503,8 +535,25 @@ class Worksheet(object):
             worksheet.update_cells(cell_list)
 
         """
-        feed = self._create_update_feed(cell_list)
-        self.client.post_cells(self, ElementTree.tostring(feed))
+
+        values_rect = cell_list_to_rect(cell_list)
+
+        start = rowcol_to_a1(cell_list[0].row, cell_list[0].col)
+        end = rowcol_to_a1(cell_list[-1].row, cell_list[-1].col)
+
+        range_label = '%s!%s:%s' % (self.title, start, end)
+
+        data = self.spreadsheet.values_update(
+            range_label,
+            params={
+                'valueInputOption': value_input_option
+            },
+            body={
+                'values': values_rect
+            }
+        )
+
+        return data
 
     def resize(self, rows=None, cols=None):
         """Resizes the worksheet.
@@ -512,44 +561,61 @@ class Worksheet(object):
         :param rows: New rows number.
         :param cols: New columns number.
         """
-        if rows is None and cols is None:
+        grid_properties = {}
+
+        if rows is not None:
+            grid_properties['rowCount'] = rows
+
+        if cols is not None:
+            grid_properties['columnCount'] = cols
+
+        if not grid_properties:
             raise TypeError("Either 'rows' or 'cols' should be specified.")
 
-        self_uri = self._get_link('self', self._element).get('href')
-        feed = self.client.get_feed(self_uri)
-        uri = self._get_link('edit', feed).get('href')
+        fields = ','.join(
+            'gridProperties/%s' % p for p in grid_properties.keys()
+        )
 
-        if rows:
-            elem = feed.find(_ns1('rowCount'))
-            elem.text = str(rows)
+        body = {
+            'requests': [{
+                'updateSheetProperties': {
+                    'properties': {
+                        'sheetId': self.id,
+                        'gridProperties': grid_properties
+                    },
+                    'fields': fields
+                }
+            }]
+        }
 
-        if cols:
-            elem = feed.find(_ns1('colCount'))
-            elem.text = str(cols)
-
-        # Send request and store result
-        self._element = self.client.put_feed(uri, ElementTree.tostring(feed))
+        return self.spreadsheet.batch_update(body)
 
     def update_title(self, title):
         """Renames the worksheet.
 
         :param title: A new title.
+
         """
 
-        self_uri = self._get_link('self', self._element).get('href')
-        feed = self.client.get_feed(self_uri)
-        uri = self._get_link('edit', feed).get('href')
+        body = {
+            'requests': [{
+                'updateSheetProperties': {
+                    'properties': {
+                        'sheetId': self.id,
+                        'title': title
+                    },
+                    'fields': 'title'
+                }
+            }]
+        }
 
-        elem = feed.find(_ns('title'))
-        elem.text = title
-
-        # Send request and store result
-        self._element = self.client.put_feed(uri, ElementTree.tostring(feed))
+        return self.spreadsheet.batch_update(body)
 
     def add_rows(self, rows):
         """Adds rows to worksheet.
 
         :param rows: Rows number to add.
+
         """
         self.resize(rows=self.row_count + rows)
 
@@ -557,95 +623,114 @@ class Worksheet(object):
         """Adds colums to worksheet.
 
         :param cols: Columns number to add.
+
         """
         self.resize(cols=self.col_count + cols)
 
-    def append_row(self, values):
+    def append_row(self, values, value_input_option='RAW'):
         """Adds a row to the worksheet and populates it with values.
         Widens the worksheet if there are more values than columns.
 
-        Note that a new Google Sheet has 100 or 1000 rows by default. You
-        may need to scroll down to find the new row.
-
         :param values: List of values for the new row.
+
         """
-        self.add_rows(1)
-        new_row = self.row_count
-        data_width = len(values)
-        if self.col_count < data_width:
-            self.resize(cols=data_width)
+        params = {
+            'valueInputOption': value_input_option
+        }
 
-        cell_list = []
-        for i, value in enumerate(values, start=1):
-            cell = self.cell(new_row, i)
-            cell.value = value
-            cell_list.append(cell)
+        body = {
+            'values': [values]
+        }
 
-        self.update_cells(cell_list)
+        return self.spreadsheet.values_append(self.title, params, body)
 
-    def insert_row(self, values, index=1):
-        """"Adds a row to the worksheet at the specified index
+    def insert_row(
+        self,
+        values,
+        index=1,
+        value_input_option='RAW'
+    ):
+        """Adds a row to the worksheet at the specified index
         and populates it with values.
 
         Widens the worksheet if there are more values than columns.
 
         :param values: List of values for the new row.
+        :param value_input_option: Determines how input data should be
+                                   interpreted. See `ValueInputOption`_
+                                   in the Sheets API.
+
+        .. _ValueInputOption: https://developers.google.com/sheets/api/reference/rest/v4/ValueInputOption
+
         """
-        if index == self.row_count + 1:
-            return self.append_row(values)
-        elif index > self.row_count + 1:
-            raise IndexError('Row index out of range')
 
-        self.add_rows(1)
-        data_width = len(values)
-        if self.col_count < data_width:
-            self.resize(cols=data_width)
+        body = {
+            "requests": [{
+                "insertDimension": {
+                    "range": {
+                      "sheetId": self.id,
+                      "dimension": "ROWS",
+                      "startIndex": index - 1,
+                      "endIndex": index
+                    }
+                }
+            }]
+        }
 
-        # Retrieve all Cells at or below `index` using a single batch query
-        top_left = rowcol_to_a1(index, 1)
-        bottom_right = rowcol_to_a1(self.row_count, self.col_count)
-        range_str = '%s:%s' % (top_left, bottom_right)
+        self.spreadsheet.batch_update(body)
 
-        cells_after_insert = self.range(range_str)
+        range_label = '%s!%s' % (self.title, 'A%s' % index)
 
-        for ind, cell in reversed(list(enumerate(cells_after_insert))):
-            if ind < self.col_count:
-                # For the first row, take the cell values from `values`
-                new_val = values[ind] if ind < len(values) else ''
-            else:
-                # For all other rows, take the cell values from the row above
-                new_val = cells_after_insert[ind - self.col_count].input_value
-            cell.value = new_val
-
-        self.update_cells(cells_after_insert)
-
-    def delete_row(self, index):
-        """"Deletes a row from the worksheet at the specified index
-
-        :param index: Index of a row for deletion
-        """
-        if index < 1 or index > self.row_count:
-            raise IndexError('Row index out of range')
-
-        # Retrieve all Cells at or below `index` using a single batch query
-        cells_after_delete = self.range(
-            index, 1, self.row_count, self.col_count
+        data = self.spreadsheet.values_update(
+            range_label,
+            params={
+                'valueInputOption': value_input_option
+            },
+            body={
+                'values': [values]
+            }
         )
 
-        # Shift rows up
-        for ind, cell in enumerate(cells_after_delete):
-            if ind + self.col_count >= len(cells_after_delete):
-                break
-            new_val = cells_after_delete[ind + self.col_count].input_value
-            cell.value = new_val
+        return data
 
-        self.update_cells(cells_after_delete)
+    def delete_row(self, index):
+        """"Deletes a row from the worksheet at the specified index.
 
-        # Remove last row
-        self.resize(rows=self.row_count - 1)
+        :param index: Index of a row for deletion.
+        """
+        body = {
+            "requests": [{
+                "deleteDimension": {
+                    "range": {
+                      "sheetId": self.id,
+                      "dimension": "ROWS",
+                      "startIndex": index - 1,
+                      "endIndex": index
+                    }
+                }
+            }]
+        }
+
+        return self.spreadsheet.batch_update(body)
+
+    def clear(self):
+        """Clears all cells in the worksheet.
+        """
+        return self.spreadsheet.values_clear(self.title)
 
     def _finder(self, func, query):
-        cells = self._fetch_cells()
+        data = self.spreadsheet.values_get(self.title)
+
+        try:
+            values = fill_gaps(data['values'])
+        except KeyError:
+            values = []
+
+        cells = [
+            Cell(row=i + 1, col=j + 1, value=value)
+            for i, row in enumerate(values)
+            for j, value in enumerate(row)
+        ]
 
         if isinstance(query, basestring):
             match = lambda x: x.value == query
@@ -671,51 +756,48 @@ class Worksheet(object):
         """
         return list(self._finder(filter, query))
 
-    def export(self, format='csv'):
-        """Export the worksheet in specified format.
-
-        :param format: A format of the output.
+    def export(self, format):
+        """.. deprecated:: 2.0
+        This feature is not supported in Sheets API v4.
         """
-        export_link = self._get_link(
-            'http://schemas.google.com/spreadsheets/2006#exportcsv',
-            self._element).get('href')
-
-        url, qs = export_link.split('?')
-        params = dict(param.split('=') for param in qs.split('&'))
-
-        params['format'] = format
-
-        params = urlencode(params)
-        export_link = '%s?%s' % (url, params)
-
-        return self.client.session.get(export_link).content
-
-    def clear(self):
-        """Clears all cells in the worksheet.
-        """
-        cells = self.range(1, 1, self.row_count, self.col_count)
-        for cell in cells:
-            cell.value = ''
-        self.update_cells(cells)
+        import warnings
+        warnings.warn(
+            "Worksheet.export() is deprecated, "
+            "this feature is not supported in Sheets API v4",
+            DeprecationWarning
+        )
 
 
 class Cell(BaseCell):
     """An instance of this class represents a single cell
-    in a :class:`worksheet <Worksheet>`.
+    in a :class:`worksheet <gspread.v4.models.Worksheet>`.
 
     """
 
-    def __init__(self, worksheet, element):
-        self.element = element
-        cell_elem = element.find(_ns1('cell'))
-        self._row = int(cell_elem.get('row'))
-        self._col = int(cell_elem.get('col'))
-
-        #: Raw value of the cell (e.g. formula)
-        self.input_value = cell_elem.get('inputValue')
-
-        numeric_value = cell_elem.get('numericValue')
-        self.numeric_value = float(numeric_value) if numeric_value else None
+    def __init__(self, row, col, value=''):
+        self._row = row
+        self._col = col
 
         #: Value of the cell.
-        self.value = cell_elem.text or ''
+        self.value = value
+
+    @property
+    def numeric_value(self):
+        try:
+            return float(self.value)
+        except ValueError:
+            return None
+
+    @property
+    def input_value(self):
+        """.. deprecated:: 2.0
+        This feature is not supported in Sheets API v4.
+        """
+        import warnings
+        warnings.warn(
+            "Cell.input_value is deprecated, "
+            "this feature is not supported in Sheets API v4. "
+            "Please use `value_render_option` when you "
+            "Retrieve `Cell` objects (e.g. in `Worksheet.range()` method).",
+            DeprecationWarning
+        )
